requirements.txt
openai>=1.0.0,<2.0.0
requests>=2.25.0,<3.0.0
python-dotenv>=0.19.0,<2.0.0
pydantic>=2.0.0,<3.0.0
click>=8.0.0,<9.0.0
asyncio-throttle>=1.0.0,<2.0.0
markdown>=3.3.0,<4.0.0
aiohttp>=3.8.0,<4.0.0
google-generativeai>=0.3.0,<1.0.0
tiktoken>=0.1.0,<1.0.0